hamming

+++ evaluation/evaluation.py
@@ -21,7 +21,7 @@ class Evaluation:
             for line in f.readlines():
                 tokens = line.strip().split()
                 if self.classification_method == "svm":
-                    node2embedding[tokens[0]] = [long(value) for value in tokens[1:]]
+                    node2embedding[tokens[0]] = [int(value) for value in tokens[1:]]
                     print(node2embedding[tokens[0]])
                 else:
                     node2embedding[tokens[0]] = [float(value) for value in tokens[1:]]
